+++ pages/student/training/[id].vue
@@ -1,0 +1,16 @@
+<template>
+  <dashboardsStudent-dashboard nav="training">
+    <div class="py-6">
+      <div class="max-w-7xl mx-auto px-4 sm:px-6 md:px-8">
+        <h1 class="text-2xl font-semibold text-gray-900">
+          Training results
+        </h1>
+      </div>
+      <div class="max-w-7xl mx-auto px-4 sm:px-6 md:px-8" />
+    </div>
+  </dashboardsStudent-dashboard>
+</template>
+
+<script setup>
+// const date = new Date().toLocaleDateString("de-DE");
+</script>